+++ test/aws/rails/sqs_active_job/configuration_test.rb
@@ -36,6 +36,17 @@ module Aws
           expect(cfg.to_h).to include(expected)
         end
 
+        # For Ruby 3.1+, Psych 4 will normally raise BadAlias error
+        it 'accepts YAML config with alias' do
+          allow_any_instance_of(ERB).to receive(:result).and_return(<<~YAML)
+common: &common
+  default: 'https://queue-url'
+queues:
+  <<: *common
+          YAML
+          expect { Aws::Rails::SqsActiveJob::Configuration.new }.to_not raise_error
+        end
+
         describe '.config' do
           before { Aws::Rails::SqsActiveJob.instance_variable_set(:@config, nil) }